// react/src/services/hubSectionStore/HubSection.type.d.ts
/*
 * Copyright The Microcks Authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *  http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type HubSectionCategory = {
  id: string;
  name: string;
};

export type HubSectionProvider = {
  id: string;
  name: string;
  count: number;
};

export type HubSectionListItem = {
  id: string;
  name: string;
  provider: string;
  providerId: string;
  description: string;
  logoUrl: string;
  categoryId: string;
};

export type HubSectionFiltersState = {
  categories: HubSectionCategory[];
  providers: HubSectionProvider[];
  selectedCategories: string[];
  selectedProviders: string[];
};

export type HubSectionFilterName = 'selectedCategories' | 'selectedProviders';

export type HubSectionListState = {
  items: HubSectionListItem[];
  packagesCount: number;
  apisCount: number;
  isLoading: boolean;
  hasError: boolean;
  error: string | null;
};